+++ pyriemann/geometry/distance.py
@@ -612,13 +612,15 @@ def distance_wasserstein(A, B, squared=False):
 
     References
     ----------
-    .. [1] `Optimal transport: old and new
-        <https://link.springer.com/book/10.1007/978-3-540-71050-9>`_
-        C. Villani. Springer Science & Business Media, 2008, vol. 338
-    .. [2] `An extension of Kakutani's theorem on infinite product measures to
+    .. [1] `An extension of Kakutani's theorem on infinite product measures to
         the tensor product of semifinite w*-algebras
         <https://www.ams.org/journals/tran/1969-135-00/S0002-9947-1969-0236719-2/S0002-9947-1969-0236719-2.pdf>`_
         D. Bures. Trans Am Math Soc, 1969, 135, pp. 199-212
+    .. [2] `On the Bures–Wasserstein distance between positive definite
+        matrices
+        <https://www.sciencedirect.com/science/article/pii/S0723086918300021>`_
+        R. Bhatia, T. Jain, Y. Lim. Expositiones mathematicae, 2019, 37,
+        pp. 165-191.
     """  # noqa
     xp = check_matrix_pair(A, B)
     B12 = sqrtm(B)